add Upstart 1.8 file
:Release Notes:
- Move the Upstart 1.8 job file from starfish-initscripts.
- Move and rename node_fork_server.upstart to
  files/launch/node_fork_server and clean out a bit of old code.
- Update CMakeFiles.txt for these changes and bump version to 2.0.1

:Detailed Notes:
The primary intent of this fix is moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and is to
move the 'imported' scripts back to their home package as soon as is
reasonable.
Additionally, there was a block in the old upstart script that tested
for more or less than 256MB RAM.  As it is unlikely for us to work on
any TV with less than 256MB, I removed the if block.

:Testing Performed:
Since the 1.8 file itself is functionally unchanged from that present
in starfish-initscripts, the testing done running Upstart 1.8 + s-i
over the last couple of months still applies.  A quick test of the
0.3.11 script on jenkins 182 showed no change from the unmodified
version of that script.

The testing done specifically for this fix is to locally build the
component and inspect the output directory:
starfish/BUILD-goldfinger/work/armv7a-vfp-neon-starfish-linux-gnueabi/ \
   mojoservicelauncher/2.0.0-71-r4/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

:QA Notes:

:Issues Addressed:
[GF-8025] add Upstart 1.8 job file to mojoservicelauncher

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: Ib2b7fceae281b0c4cb336aefaaa1f22887fd711c
Reviewed-on: https://g2g.palm.com/2773
Reviewed-by: Build Verification
Reviewed-by: Mark Bessey <mark.bessey@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Steve Lemke <steve.lemke@lge.com>