Pass additional command-line arguments to services
:Release Notes:
The run-js-service script was failing to pass additional command line
parameters after the -- to the executing service. This is a useful way
to allow services to tune their behavior without having to modify the source
Also, changed the order that the bootstrap code looks for files in order to
determine whether a service uses the old mojoservice framework or the newer
webos-service framework.

:Detailed Notes:

:Testing Performed:
Tested by running the devmode service in "test mode" (no reboots) by
passing the --testmode option to run-js-service
:QA Notes:

:Issues Addressed:
[GF-44509] run-js-service should pass additional command-line arguments to services
[GF-44518] run-js-service should prefer loading from package.json, when available

Open-webOS-DCO-1.0-Signed-off-by: Mark Bessey <mark.bessey@lge.com>

Change-Id: I54b4bd5952d7b02a98f443c45101794cf3629973
Reviewed-on: https://g2g.palm.com/3779
Reviewed-by: Build Verification
Reviewed-by: Mark Bessey <mark.bessey@lge.com>
Tested-by: Mark Bessey <mark.bessey@lge.com>