Configure node-inspector based on file entries
:Release Notes:
Node services can enable and configure the debugger through files in
the service directory, for Developer Mode services.

:Detailed Notes:
When running in developer mode, we needed a way to enable/disable the
node-inspector (and configure the ports) without being able to modify
the command-line arguments for running the service. We now pull that
configuration from the files $SERVICE_PATH/_ares/debugger_port and
$SERVICE_PATH/_ares/inspector_port

:Testing Performed:
Ran a service with and without both files, and with the files empty
and with contents. If the file has contents, those are used to set
the port. If it's empty, then the defaults are used.

:QA Notes:

:Issues Addressed:
[GF-41702] Fix ares-inspect so it can launch service for debugging from devmode jail

Open-webOS-DCO-1.0-Signed-off-by: Mark Bessey <mark.bessey@lge.com>

Change-Id: Ia558e0c667c57b75416f65c173f5c5da62280096
Reviewed-on: https://g2g.palm.com/3946
Reviewed-by: DCO Verification
Tested-by: Mark Bessey <mark.bessey@lge.com>
Reviewed-by: Mark Bessey <mark.bessey@lge.com>